# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/discount": {
            "post": {
                "description": "return usage discounts for a giftCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Discount"
                ],
                "summary": "Get discounts by giftCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gift_code",
                        "name": "gift_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/giftCharge": {
            "get": {
                "description": "Get giftCharge list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiftCharge"
                ],
                "summary": "Get giftCharge list information",
                "parameters": [
                    {
                        "enum": [
                            true,
                            false
                        ],
                        "type": "boolean",
                        "description": "giftCharge status",
                        "name": "isActive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.GiftCharge"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create new giftcharge and send it",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiftCharge"
                ],
                "summary": "Create new giftcharge",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GiftCharge"
                        }
                    }
                }
            }
        },
        "/giftCharge/{gift_id}": {
            "get": {
                "description": "Get giftCharge by giftCharge uuid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GiftCharge"
                ],
                "summary": "Get giftCharge by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gift_id",
                        "name": "gift_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GiftCharge"
                        }
                    }
                }
            }
        },
        "/report/{gift_code}": {
            "get": {
                "description": "return usage reports for a giftCode",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get reports by giftCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gift_code",
                        "name": "gift_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Report"
                            }
                        }
                    }
                }
            }
        },
        "/report/{mobile}": {
            "get": {
                "description": "return usage reports for a mobile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Report"
                ],
                "summary": "Get reports by mobile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "gift_code",
                        "name": "gift_code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Report"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.GiftCharge": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "maxUsageCount": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "validityPeriodEnd": {
                    "type": "string"
                },
                "validityPeriodStart": {
                    "type": "string"
                }
            }
        },
        "models.Report": {
            "type": "object",
            "properties": {
                "chargeAmount": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "giftCode": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mobile": {
                    "type": "string"
                },
                "reportTime": {
                    "type": "string"
                }
            }
        }
    }
}